create abstract folder include model, serializer and viewset for avoid DRY
