+++ core/abstract/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+class AbstractSerlizer(serializers.ModelSerializer):
+    id = serializers.UUIDField(source="public_id", read_only=True, format="hex")
+    created = serializers.DateTimeField(read_only=True)
+    updated = serializers.DateTimeField(read_only=True)
+    
+
