+++ core/abstract/viewset.py
@@ -1,0 +1,6 @@
+from rest_framework.viewsets import ModelViewSet
+from rest_framework import filters
+class AbstractViewSet(ModelViewSet):
+    filter_backends = [filters.OrderingFilter]
+    ordering_fields = ['updated', 'created']
+    ordering = ['-updated']
