+++ core/user/serializers.py
@@ -1,9 +1,6 @@
-from rest_framework import serializers
+from core.abstract.serializers import AbstractSerlizer
 from .models import User
-class UserSerializer(serializers.ModelSerializer):
-    id = serializers.UUIDField(source='public_id', read_only=True, format='hex')
-    created = serializers.DateTimeField(read_only=True)
-    updated = serializers.DateTimeField(read_only=True)
+class UserSerializer(AbstractSerlizer):
     
     class Meta:
         model = User
